% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomeStation.R
\name{createHomeCareSiteSettings}
\alias{createHomeCareSiteSettings}
\title{Creates settings object}
\usage{
createHomeCareSiteSettings(
  connectionDetails = NULL,
  con = NULL,
  cohortSchema,
  cohortTable,
  cdmSchema,
  method = NULL,
  careSiteColumn = "care_site_name"
)
}
\arguments{
\item{connectionDetails}{connection details for DatabaseConnector}

\item{con}{database connection}

\item{cohortSchema}{schema containing the cohort table}

\item{cohortTable}{name of the cohort table}

\item{cdmSchema}{schema of the OMOP CDM}

\item{method}{character of length 1, how to derive home care site ("mostRecent", "mostFrequent")}

\item{careSiteColumn}{name of column in care_site table to use, default is standard "care_site_name"}
}
\value{
list of class "homeCareSiteSettings"
}
\description{
Creates settings object
}
