% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomeStation.R
\name{deriveHomeCareSite}
\alias{deriveHomeCareSite}
\title{Derive home care site}
\usage{
deriveHomeCareSite(settings)
}
\arguments{
\item{settings}{object of class "homeCareSiteSettings"}
}
\value{
list of class "homeSiteServerTable"
}
\description{
Derive home care site
}
