% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomeStation.R
\name{downloadHomeCareSite}
\alias{downloadHomeCareSite}
\title{Download derived home care site table from server}
\usage{
downloadHomeCareSite(tableReference)
}
\arguments{
\item{settings}{object of class "homeCareSiteServerTable"}
}
\value{
data frame
}
\description{
Download derived home care site table from server
}
