% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomeStation.R
\name{summary.homeCareSiteServerTable}
\alias{summary.homeCareSiteServerTable}
\title{Summary generic method for class "homeCareSiteServerTable"}
\usage{
\method{summary}{homeCareSiteServerTable}(tableReference)
}
\arguments{
\item{tableReference}{list of class "homeCareSiteServerTable"}
}
\value{
list of summary data
}
\description{
Summary generic method for class "homeCareSiteServerTable"
}
